moved tests around. Implemented composer editor tests

+++ src/Factories/ModuleMigrationFactory.php
@@ -25,7 +25,9 @@ class ModuleMigrationFactory extends FileFactory implements Contract
      */
     public function undo(): void
     {
-        $this->fileSystem->delete(base_path($this->getRelativeMigrationDir()) . "/" . $this->getMigrationName());
+        if ($this->fileSystem->isFile(base_path($this->getRelativeMigrationDir()) . "/" . $this->getMigrationName())) {
+            $this->fileSystem->delete(base_path($this->getRelativeMigrationDir()) . "/" . $this->getMigrationName());
+        }
     }
 
     /**
